--- agriculture_app/README.md
@@ -1,3 +1,0 @@
-# AgricultureApp
-
-Primer Commit
