--- agriculture_app/config.init
@@ -1,2 +1,0 @@
-#Sat Feb 08 19:58:30 COT 2020
-Language=language/languageES.properties
