--- agriculture_app/language/languageES.properties
@@ -1,40 +1,0 @@
-#FramePrin
-App_Name = AGRICAPP®
-
-
-Language = Idioma
-Language_ES = Español
-Language_US = Inglés
-
-Home = Inicio
-Sign_up = Registrarse
-Login = Ingresar
-Gest_dt = Administrar datos
-
-#Table
-Crop = Cultivo 
-Harvested_area = Area Cosechada
-Planted_area = Area Sembrada
-Type_al = Tipo producto
-Price = Precio(Ton)
-Costs = Costos(Ha)
-Tons = Toneladas
-
-Data = Datos
-Reports = Reportes
-
-Contact = Contactenos
-Contact_info = Dirección: Carrera 5 No. 10-125 Tunja - Boyacá - Colombia\nLíneas telefónicas: +57(8) 7502189 - 7501951\nFax: +57(8) 7500770\nLínea Gratuita: 018000918791\nE-mail: contactenosagricapp@gmail.com
-
-Privacy = Política de privacidad y uso de la app
-Who = ¿Quienes somos?
-About = Acerca de
-Help = Ayuda
-Policies = Politicas
-
-Add_crop = Agregar cultivo
-Del_crop = Eliminar cultivo
-Edit_crop = Modificar cultivo
-
-Copyright = © Derechos de Autor 2020 por UPTC. Reservados todos los derechos. 
-
