--- agriculture_app/language/languageUS.properties
@@ -1,39 +1,0 @@
-#FramePrin
-App_Name = AGRICAPP®
-
-
-Language = Language
-Language_ES = Spanish
-Language_US = English
-
-Home = Home
-Sign_up = Sign Up
-Login = Login
-Gest_dt = Manage dates
-
-#Table
-Crop = Crop 
-Harvested_area = Harvested area
-Planted_area = Planted area
-Type_al = Type product
-Price = Price(Ton)
-Costs = Costs(Ha)
-Tons = Tons
-
-Data = Data
-Reports = Reports
-
-Contact = Contact us
-Contact_info = Adress: Career 5 No. 10-125 Tunja - Boyacá - Colombia\nTelephone Lines: +57(8) 7502189 - 7501951\nFax: +57(8) 7500770\nFree Line: 018000918791\nE-mail: contactenosagricapp@gmail.com
-
-Privacy = Privacy policy and use of the app
-Who = About us?
-About = About
-Help = Help
-Policies = Policies
-
-Add_crop = Add crop
-Del_crop = Remove crop
-Edit_crop = Modify crop
-
-Copyright = © Copyright 2020 by UPTC. All rights reserved.
