--- agriculture_app/src/constants/ConstantsLanguage.java
@@ -1,49 +1,0 @@
-package constants;
-
-public class ConstantsLanguage {
-	
-	public static final String APP_NAME = "App_Name";
-	public static final String LANG = "Language";
-	public static final String LANG_ES = "Language_ES";
-	public static final String LANG_US = "Language_US";
-	public static final String SIGN_UP = "Sign_up";
-	public static final String LOGIN = "Login";
-	
-	public static final String NAME_CROOP = "Nombre del cultivo";
-	public static final String TYPE_CROOP = "Tipo de cultivo";
-	public static final String PRICE_CROOP = "Precio del cultivo";
-	public static final String HARVESTED_AREA = "Area Cosechada";
-	public static final String SOWN_AREA = "Area Sembrada";
-	public static final String COST_SALE = "Valor de venta";
-	public static final String TONS_ADD = "Toneladas";
-	
-	public static final String CREATE_CROP = "Add_crop";
-	public static final String ADD_CROP = "Agregar cultivo";
-	public static final String CANCEL = "Cancelar";
-	public static final String DEL_CROP = "Del_crop";
-	public static final String EDIT_CROP = "Edit_crop";
-	
-	public static final String HOME = "Home";
-	public static final String DATA = "Data";
-	public static final String REPORTS = "Reports";
-	public static final String GEST_DT = "Gest_dt";
-	
-	//Table
-	public static final String HARV_AREA = "Harvested_area";
-	public static final String PRICE = "Price";
-	public static final String TYPE_AL = "Type_al";
-	public static final String CROP = "Crop";
-	public static final String PLANT_AREA = "Planted_area";
-	public static final String TONS = "Tons";
-	public static final String COSTS = "Costs";
-	
-	//Footer
-	public static final String CONTACT = "Contact";
-	public static final String PRIVACY = "Privacy";
-	public static final String WHO = "Who";
-	public static final String ABOUT = "About";
-	public static final String HELP = "Help";
-	public static final String POLICIES = "Policies";
-	public static final String CONTACT_INFO = "Contact_info";
-	public static final String COPYRIGHT = "Copyright";
-}
