--- agriculture_app/src/constants/ConstantsView.java
@@ -1,31 +1,0 @@
-package constants;
-
-import java.awt.Color;
-import java.awt.Font;
-
-public class ConstantsView {
-	
-	public static final String LOGO_APP = "images/logo.png";
-	public static final String MANAGE_DATA = "images/edit.png";
-	public static final String BANNER = "images/banner.png";
-	public static final String DATA = "images/data.png";
-	public static final String REPORTS = "images/reports.png";
-	public static final String USER = "images/user.png";
-	public static final String CONTACT = "images/contact.png";
-	public static final String POLICIES = "images/policies.png";
-	public static final String SPAIN = "images/spain.png";
-	public static final String ENG = "images/eng.png";
-	public static final String HOME = "images/home.png";
-	public static final String ADD = "images/add.png";
-	public static final String CENTER = "images/head north.jpg";
-	public static final String MODIFY = "images/modify.png";
-	public static final String DELETE = "images/delete.png";
-	public static final String NULL = " ";
-	
-	public static final Color GREEN = new Color(203, 229, 88);
-	public static final Color ORANGE = new Color(255, 203, 124);
-	public static final Color GREEN_FOOTER = new Color(115, 134, 116);
-	public static final Font BAUHAUS_15 = new Font("Bauhaus 93", Font.PLAIN, 15);
-	public static final Font BAUHAUS_17_PLAIN = new Font("Bauhaus 93", Font.PLAIN, 17);
-	public static final Font BAUHAUS_17 = new Font("Bauhaus 93", Font.BOLD, 17);
-}
