--- agriculture_app/src/controller/Command.java
@@ -1,8 +1,0 @@
-package controller;
-
-public enum Command {
-	
-	LANGUAGE_ES, LANGUAGE_ENG, 
-	DATA, MANAGE_DATA, HOME, REPORTS,ADD_CROOP,CREATE_CROOP, CANCEL;
-
-}
